Move InputMap to Common for server to reference
