+++ Phoenix/Common/Include/Common/CMakeLists.txt
@@ -3,20 +3,21 @@ add_subdirectory(Voxels)
 
 set(currentDir ${CMAKE_CURRENT_LIST_DIR})
 set(Headers
-	${mathHeaders}
-	${voxelHeaders}
+		${mathHeaders}
+		${voxelHeaders}
 
-	${currentDir}/CoreIntrinsics.hpp
-	${currentDir}/EnumTools.hpp
-	${currentDir}/Singleton.hpp
-	${currentDir}/FileIO.hpp
-	${currentDir}/Logger.hpp
+		${currentDir}/CoreIntrinsics.hpp
+		${currentDir}/EnumTools.hpp
+		${currentDir}/Singleton.hpp
+		${currentDir}/FileIO.hpp
+		${currentDir}/Logger.hpp
 
-	${currentDir}/ContentLoader.hpp
-	${currentDir}/Settings.hpp
-	${currentDir}/Commander.hpp
-	${currentDir}/Position.hpp
-	${currentDir}/Movement.hpp
+		${currentDir}/ContentLoader.hpp
+		${currentDir}/Settings.hpp
+		${currentDir}/InputMap.hpp
+		${currentDir}/Commander.hpp
+		${currentDir}/Position.hpp
+		${currentDir}/Movement.hpp
 
-	PARENT_SCOPE
-)
+		PARENT_SCOPE
+		)
