+++ Phoenix/Common/Source/CMakeLists.txt
@@ -3,13 +3,14 @@ add_subdirectory(Voxels)
 
 set(currentDir ${CMAKE_CURRENT_LIST_DIR})
 set(Sources
-	${mathSources}
-	${voxelSources}
+		${mathSources}
+		${voxelSources}
 
-	${currentDir}/ContentLoader.cpp
-	${currentDir}/Settings.cpp
-	${currentDir}/Logger.cpp
-	${currentDir}/Commander.cpp
+		${currentDir}/ContentLoader.cpp
+		${currentDir}/Settings.cpp
+		${currentDir}/Logger.cpp
+		${currentDir}/Commander.cpp
+		${currentDir}/InputMap.cpp
 
-	PARENT_SCOPE
-)
+		PARENT_SCOPE
+		)
